requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
idna==2.9
pathspec==0.7.0
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
